extract tasks
